Пререписывает withAddComment c HOC на hook
